feat: support ihostbuilder as well, and package updates
